requirements.txt
# Vision
ultralytics
transformers
pillow
opencv-python

# RAG
chromadb
pdfplumber
python-docx

# LLM / Embeddings (via Ollama)
ollama

# Schemas & Settings
pydantic>=2.0
pydantic-settings

# Logging
structlog

# Tests
pytest
pytest-cov

# Fine-tuning
datasets>=2.18
accelerate>=0.27

# Utilities
tqdm
